# docs/openapi.json
{
  "components": {
    "schemas": {
      "ActivityOut": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "person_name": {
            "title": "Person Name",
            "type": "string"
          },
          "recognized": {
            "title": "Recognized",
            "type": "boolean"
          }
        },
        "required": [
          "created_at",
          "person_id",
          "person_name",
          "recognized",
          "confidence"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
...
      "Body_landmarks_api_landmarks_image_post": {
        "properties": {
          "image": {
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "type": "string"
          },
          "include_points": {
            "default": true,
            "description": "Incluir los 478 puntos de cada rostro",
            "title": "Include Points",
            "type": "boolean"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_landmarks_api_landmarks_image_post",
        "type": "object"
      },
...
      "BoundingBoxOut": {
        "properties": {
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBoxOut",
        "type": "object"
      },
...
      "DailyCountOut": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "exits": {
            "title": "Exits",
            "type": "integer"
          },
          "recognized": {
            "title": "Recognized",
            "type": "integer"
          },
          "unknown": {
            "title": "Unknown",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "recognized",
          "unknown",
          "entries",
          "exits"
        ],
        "title": "DailyCountOut",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "EyesOut": {
        "properties": {
          "left_blink_score": {
            "title": "Left Blink Score",
            "type": "number"
          },
          "left_ear": {
            "title": "Left Ear",
            "type": "number"
          },
          "right_blink_score": {
            "title": "Right Blink Score",
            "type": "number"
          },
          "right_ear": {
            "title": "Right Ear",
            "type": "number"
          }
        },
        "required": [
          "left_ear",
          "right_ear",
          "left_blink_score",
          "right_blink_score"
        ],
        "title": "EyesOut",
        "type": "object"
      },
...
      "FaceLandmarksOut": {
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BoundingBoxOut"
          },
          "eyes": {
            "$ref": "#/components/schemas/EyesOut"
          },
          "head_pose": {
            "$ref": "#/components/schemas/HeadPoseOut"
          },
          "points": {
            "default": [],
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "bbox",
          "head_pose",
          "eyes"
        ],
        "title": "FaceLandmarksOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeadPoseOut": {
        "properties": {
          "pitch": {
            "description": "Grados; > 0 = la persona mira hacia arriba",
            "title": "Pitch",
            "type": "number"
          },
          "roll": {
            "description": "Grados; > 0 = rotación antihoraria en la imagen",
            "title": "Roll",
            "type": "number"
          },
          "yaw": {
            "description": "Grados; > 0 = la persona gira hacia su derecha",
            "title": "Yaw",
            "type": "number"
          }
        },
        "required": [
          "yaw",
          "pitch",
          "roll"
        ],
        "title": "HeadPoseOut",
        "type": "object"
      },
...
      "LandmarksOut": {
        "properties": {
          "faces": {
            "items": {
              "$ref": "#/components/schemas/FaceLandmarksOut"
            },
            "title": "Faces",
            "type": "array"
          },
          "faces_detected": {
            "title": "Faces Detected",
            "type": "integer"
          },
          "image_height": {
            "title": "Image Height",
            "type": "integer"
          },
          "image_width": {
            "title": "Image Width",
            "type": "integer"
          }
        },
        "required": [
          "faces_detected",
          "image_width",
          "image_height",
          "faces"
        ],
        "title": "LandmarksOut",
        "type": "object"
      },
...
      "PersonCountOut": {
        "properties": {
          "average_confidence": {
            "title": "Average Confidence",
            "type": "number"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "person_name": {
            "title": "Person Name",
            "type": "string"
          },
          "recognitions": {
            "title": "Recognitions",
            "type": "integer"
          }
        },
        "required": [
          "person_id",
          "person_name",
          "recognitions",
          "average_confidence"
        ],
        "title": "PersonCountOut",
        "type": "object"
      },
...
      "StatisticsOut": {
        "properties": {
          "average_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Confidence"
          },
          "by_day": {
            "items": {
              "$ref": "#/components/schemas/DailyCountOut"
            },
            "title": "By Day",
            "type": "array"
          },
          "by_person": {
            "items": {
              "$ref": "#/components/schemas/PersonCountOut"
            },
            "title": "By Person",
            "type": "array"
          },
          "date_from": {
            "format": "date",
            "title": "Date From",
            "type": "string"
          },
          "date_to": {
            "format": "date",
            "title": "Date To",
            "type": "string"
          },
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "exits": {
            "title": "Exits",
            "type": "integer"
          },
          "recognized_total": {
            "title": "Recognized Total",
            "type": "integer"
          },
          "unknown_total": {
            "title": "Unknown Total",
            "type": "integer"
          }
        },
        "required": [
          "date_from",
          "date_to",
          "recognized_total",
          "unknown_total",
          "entries",
          "exits",
          "average_confidence",
          "by_day",
          "by_person"
        ],
        "title": "StatisticsOut",
        "type": "object"
      },
      "SummaryOut": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "people_present": {
            "title": "People Present",
            "type": "integer"
          },
          "persons_registered": {
            "title": "Persons Registered",
            "type": "integer"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/ActivityOut"
            },
            "title": "Recent Activity",
            "type": "array"
          },
          "recognitions_today": {
            "title": "Recognitions Today",
            "type": "integer"
          },
          "unknown_today": {
            "title": "Unknown Today",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "persons_registered",
          "recognitions_today",
          "people_present",
          "unknown_today",
          "recent_activity"
        ],
        "title": "SummaryOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Cookie HttpOnly que entrega POST /api/auth/login.",
        "in": "cookie",
        "name": "facetrack_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nReconocimiento facial, asistencia y estadísticas con **procesamiento local**.\n\n**Autenticación.** `POST /api/auth/login` entrega una cookie de sesión HttpOnly;\nel navegador la envía sola en las demás llamadas. Todas las rutas la requieren\nsalvo `GET /api/health` y el login. Las marcadas *Solo ADMIN* exigen ese rol.\n\n**Errores.** Siempre `{\"detail\": str, \"code\": str}`; los 5xx agregan\n`request_id` (también en la cabecera `X-Request-ID` de toda respuesta), que\npermite encontrar el detalle en los logs.\n\n**Privacidad.** Las imágenes se procesan en memoria y se descartan; ninguna\nrespuesta incluye embeddings.\n",
    "title": "FaceTrack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/estadisticas": {
      "get": {
        "operationId": "statistics_api_estadisticas_get",
        "parameters": [
          {
            "description": "Por defecto, hace 6 días",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Por defecto, hace 6 días",
              "title": "From"
            }
          },
          {
            "description": "Por defecto, hoy",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Por defecto, hoy",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Statistics",
        "tags": [
          "estadisticas"
        ]
      }
    },
    "/api/estadisticas/resumen": {
      "get": {
        "description": "Números del dashboard para hoy (hora local) y la actividad reciente.",
        "operationId": "summary_api_estadisticas_resumen_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Summary",
        "tags": [
          "estadisticas"
        ]
      }
    },
...
    "/api/landmarks/image": {
      "post": {
        "description": "Malla facial (478 puntos), apertura de ojos (EAR) y orientación de la cabeza de cada rostro.",
        "operationId": "landmarks_api_landmarks_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_landmarks_api_landmarks_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LandmarksOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Imagen inválida o corrupta"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Archivo demasiado grande"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Formato no soportado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Landmarks",
        "tags": [
          "liveness"
        ]
      }
    },
...
  },
...
}